requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.0

# Google AI/ML
google-generativeai==0.8.6
google-cloud-texttospeech==2.34.0

# LLM APIs
groq==1.0.0

# RAG & Embeddings  
qdrant-client
langchain-google-genai
numpy>=1.25.0
sentence-transformers

# Speech Recognition
SpeechRecognition==3.14.5

# Utilities
requests==2.31.0
aiofiles==23.2.1
deep-translator==1.11.4
edge-tts==6.1.19
elevenlabs==1.7.0
